# apps/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ClassRoster" : {
        "properties" : {
          "lop" : {
            "$ref" : "#/components/schemas/ClassSection"
          },
          "sinhVien" : {
            "items" : {
              "$ref" : "#/components/schemas/RosterEntry"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "ClassSection" : {
        "properties" : {
          "choPhepLienCoSo" : {
            "type" : "boolean"
          },
          "hinhThucHoc" : {
            "type" : "string"
          },
          "maCoSoHost" : {
            "type" : "string"
          },
          "maGiangVien" : {
            "type" : "string"
          },
          "maHocKy" : {
            "type" : "string"
          },
          "maLopHP" : {
            "type" : "string"
          },
          "maMonHoc" : {
            "type" : "string"
          },
          "phienBanLich" : {
            "format" : "int32",
            "type" : "integer"
          },
          "soLuongDaDangKy" : {
            "format" : "int32",
            "type" : "integer"
          },
          "soLuongToiDa" : {
            "format" : "int32",
            "type" : "integer"
          },
          "soTinChi" : {
            "format" : "int32",
            "type" : "integer"
          },
          "tenGiangVien" : {
            "type" : "string"
          },
          "tenMonHoc" : {
            "type" : "string"
          },
          "trangThai" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "CourseDetail" : {
        "properties" : {
          "mon" : {
            "$ref" : "#/components/schemas/CourseSummary"
          },
          "monPhuThuoc" : {
            "items" : {
              "$ref" : "#/components/schemas/CourseSummary"
            },
            "type" : "array"
          },
          "tienQuyet" : {
            "items" : {
              "$ref" : "#/components/schemas/CourseSummary"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "CourseReport" : {
        "properties" : {
          "capNhatLuc" : {
            "format" : "date-time",
            "type" : "string"
          },
          "monHoc" : {
            "items" : {
              "$ref" : "#/components/schemas/CourseStat"
            },
            "type" : "array"
          },
          "phamVi" : {
            "$ref" : "#/components/schemas/ReportScope"
          }
        },
        "type" : "object"
      },
      "CourseStat" : {
        "properties" : {
          "chuaCoKetQua" : {
            "format" : "int32",
            "type" : "integer"
          },
          "luotDangKy" : {
            "format" : "int32",
            "type" : "integer"
          },
          "maMonHoc" : {
            "type" : "string"
          },
          "phanBoDiem" : {
            "items" : {
              "$ref" : "#/components/schemas/GradeBucket"
            },
            "type" : "array"
          },
          "soDat" : {
            "format" : "int32",
            "type" : "integer"
          },
          "soLop" : {
            "format" : "int32",
            "type" : "integer"
          },
          "soTruot" : {
            "format" : "int32",
            "type" : "integer"
          },
          "tenMonHoc" : {
            "type" : "string"
          },
          "tiLeLapDay" : {
            "type" : "number"
          },
          "tongDaDangKy" : {
            "format" : "int32",
            "type" : "integer"
          },
          "tongSucChua" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "CourseSummary" : {
        "properties" : {
          "maKhoa" : {
            "type" : "string"
          },
          "maMonHoc" : {
            "type" : "string"
          },
          "soTinChi" : {
            "format" : "int32",
            "type" : "integer"
          },
          "tenKhoa" : {
            "type" : "string"
          },
          "tenMonHoc" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "CreateClassRequest" : {
        "properties" : {
          "choPhepLienCoSo" : {
            "type" : "boolean"
          },
          "hinhThucHoc" : {
            "minLength" : 1,
            "type" : "string"
          },
          "maGiangVien" : {
            "type" : "string"
          },
          "maHocKy" : {
            "minLength" : 1,
            "type" : "string"
          },
          "maMonHoc" : {
            "minLength" : 1,
            "type" : "string"
          },
          "soLuongToiDa" : {
            "format" : "int32",
            "maximum" : 500,
            "minimum" : 1,
            "type" : "integer"
          }
        },
        "required" : [ "hinhThucHoc", "maHocKy", "maMonHoc" ],
        "type" : "object"
      },
      "CreateCourseRequest" : {
        "properties" : {
          "maKhoa" : {
            "minLength" : 1,
            "type" : "string"
          },
          "maMonHoc" : {
            "maxLength" : 20,
            "minLength" : 0,
            "type" : "string"
          },
          "soTinChi" : {
            "format" : "int32",
            "maximum" : 15,
            "minimum" : 1,
            "type" : "integer"
          },
          "tenMonHoc" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "maKhoa", "maMonHoc", "tenMonHoc" ],
        "type" : "object"
      },
...
      "GradeBucket" : {
        "properties" : {
          "khoang" : {
            "type" : "string"
          },
          "soLuong" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "RemoveEnrollmentRequest" : {
        "properties" : {
          "lyDo" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "lyDo" ],
        "type" : "object"
      },
      "ReportScope" : {
        "properties" : {
          "maCoSo" : {
            "type" : "string"
          },
          "maHocKy" : {
            "type" : "string"
          },
          "maMonHoc" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "RosterEntry" : {
        "properties" : {
          "hoTen" : {
            "type" : "string"
          },
          "maCoSoNha" : {
            "type" : "string"
          },
          "maSinhVien" : {
            "type" : "string"
          },
          "ngayDangKy" : {
            "format" : "date-time",
            "type" : "string"
          },
          "trangThai" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "SetPrerequisitesRequest" : {
        "properties" : {
          "tienQuyet" : {
            "items" : {
              "type" : "string"
            },
            "maxItems" : 20,
            "minItems" : 0,
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "UpdateCourseRequest" : {
        "properties" : {
          "maKhoa" : {
            "minLength" : 1,
            "type" : "string"
          },
          "soTinChi" : {
            "format" : "int32",
            "maximum" : 15,
            "minimum" : 1,
            "type" : "integer"
          },
          "tenMonHoc" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "maKhoa", "tenMonHoc" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "PTITONE_AT" : {
        "description" : "Cookie HttpOnly do server đặt khi đăng nhập. Khai báo ở đây để spec nói rõ API dùng cookie; không nhập tay được và cũng không cần.",
        "in" : "cookie",
        "name" : "PTITONE_AT",
        "type" : "apiKey"
      }
    }
  },
  "info" : {
    "description" : "Đăng nhập bằng `POST /api/auth/login` ngay trong trang này: cookie phiên được trình duyệt giữ, các lệnh sau tự dùng. Không có `Authorization` header.\n\nBảng quyền theo vai trò, 36 mã lỗi và các quy tắc nghiệp vụ nằm ở `docs/PTIT-One-API-Contract.md`; trang này chỉ mô tả hình dạng request/response.\n\nPhạm vi cơ sở LUÔN lấy từ JWT đã ký — không endpoint nào nhận tham số `maCoSo`.",
    "title" : "PTIT One API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/classes" : {
      "get" : {
        "operationId" : "search_3",
        "parameters" : [ {
          "in" : "query",
          "name" : "maHocKy",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "maMonHoc",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "maGiangVien",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ClassSection"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "class-section-controller" ]
      },
      "post" : {
        "operationId" : "create_5",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateClassRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClassSection"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "class-section-controller" ]
      }
    },
...
    "/api/classes/{maLopHP}/students" : {
      "get" : {
        "operationId" : "students",
        "parameters" : [ {
          "in" : "path",
          "name" : "maLopHP",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClassRoster"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "roster-controller" ]
      }
    },
    "/api/classes/{maLopHP}/students/{maSinhVien}/remove" : {
      "post" : {
        "operationId" : "removeStudent",
        "parameters" : [ {
          "in" : "path",
          "name" : "maLopHP",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "maSinhVien",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RemoveEnrollmentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClassSection"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "roster-controller" ]
      }
    },
...
    "/api/courses" : {
      "get" : {
        "operationId" : "search_2",
        "parameters" : [ {
          "in" : "query",
          "name" : "maKhoa",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CourseSummary"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "course-controller" ]
      },
      "post" : {
        "operationId" : "create_4",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateCourseRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CourseDetail"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "course-controller" ]
      }
    },
    "/api/courses/{maMonHoc}" : {
      "delete" : {
        "operationId" : "delete_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "maMonHoc",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "course-controller" ]
      },
      "get" : {
        "operationId" : "detail_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "maMonHoc",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CourseDetail"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "course-controller" ]
      },
      "put" : {
        "operationId" : "update_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "maMonHoc",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateCourseRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CourseDetail"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "course-controller" ]
      }
    },
    "/api/courses/{maMonHoc}/prerequisites" : {
      "put" : {
        "operationId" : "setPrerequisites",
        "parameters" : [ {
          "in" : "path",
          "name" : "maMonHoc",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SetPrerequisitesRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CourseDetail"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "course-controller" ]
      }
    },
...
    "/api/reports/courses" : {
      "get" : {
        "operationId" : "courses",
        "parameters" : [ {
          "in" : "query",
          "name" : "maHocKy",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "maCoSo",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CourseReport"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "report-controller" ]
      }
    },
...
  }
}